Implements most crawler functionality
Ported directly from python. The translation is relatively simple, and
the final solution is more compact. Need to test to be extra sure that
everything is working properly.

Note that this doesn't handle creating a component and interactively
using it, it simply is a series of functions to perform each task.