Fixes channel blocking bug
Due to a lack of specified buffer sizes for channels one of the channels
was blocking when attempting to do a put operation on the channel (even
though it was unbuffered). I'm not sure if this is the specified
behaviour, and it was very odd. However, giving channels a specific size
is a good practice anyways.

This also introduces a while loop that will continuously poll if we
enter the MFA code incorrectly. This is great for handling any type of
user error.