Implements full payment We can now use this code to actually pay the card. I tested it and it actually worked perfectly. It handles waiting for an MFA code, making a transaction approval request, and finally executing the payment.
We still need a sanity check to make sure it works. What I should do is
take a look at the HTML for the success payment and use that as
verification, which I can then sound out to a potential listener.

I also need to refactor this and clean it up. The main logic is
extremely messy and needs to be separated out. We also should emulate
and carefully manage the state so that we can write unit tests without
having to actually hit all of these weird edge cases (like using
different computers and networks to try and actually get an MFA code).