Adds outputs to all actions
Instead of actions just doing things, now all of them will report
whether they succeed, or move to the next phase in something. This is
needed because we have to plug some sort of response into our virtual
state to move it forward. As of now many of these are actually
incomplete and should be improved by explicity checking that things are
a success (as opposed to simply assuming), and returning failure if it
occurs. Literally all of these things can fail but we don't really check
for that and it's problematic.

However, that's just a sign of our new design informing us of a problem.
It's good that we're catching it.