Fully rewrites all of the state related functions
Makes it so that we have a full set of functions that all are capable of
handling the action format, and will only be executed in the proper
order. There's definitely room for improvement, but this basic intial
design should work. The next step is testing all of this virtually.