Fixes many errors in the crawler state, adds test
Made the first successful running test for the program, that effectively
simulates the state and steps through the different state transitions.
Along the way this led to many alterations of the source, most notably
the short circuit evaluation of errors (so that errors are not
overwritten and therefore much more addressable).

As I write this code I can see very easy that atoms, while great for
prototyping do not scale, and that we should namespace these different
updates and state transitions quickly so that referencing doesn't become
an issue, or can at least be managed entirely by the
interpreter/compiler.