Fixes issue with case statement
The reason that the case statement was failing is that case takes a
literal string or some other type of constant within it. You can't use
references! I almost think that this is a bug, but there's probably a
good reason for it.

Anyways, I fixed this issue by instead using a `condp` statement. This
does what we'd expect and this was validated in the REPL.