We can now successfully make a test payment
Finally after rearchitecting everything we can make payments sucessfully
by carefully threading state. However this only works in the case where
we don't have to deal with MFA, that has yet to be tested. We need to
make sure that it works in the headless browser to be totally sure that
everything is good.