Improves the interface of the CLI Makes it so that the crawler can interact much better over the CLI and adds an interactive component.
However I notice that there
would be a great overall improvement if I start to use the specs. It
would help to reduce redundancy and prevent errors all throughout the
program. I just ran into a bug where I forgot to set the variables for
credentials on the command line and because of that I got some sort of
odd bug, when that would have easily been caught had I created a sepc
for what credentials should looks like.