Starts utilizing Spec
It turns out that `atoms` have a very handle validator function that we
can use to make sure that our state stays valid from update to update.
Thanks to this it means that we just need to create a spec for our state
object and pass that as the validator. Everything then gets taken care
of for us!

In addition, spec comes with some extremely powerful debugging tools
that help us to describe any errors. So we know exactly why something
fails to satisfy the spec, and how to address it. It's like I have the
tight control and safety of Haskell, but everything is still flexible.